Transposes added
PCA and PLS often want data as column vectors, but it is much cleaner to
store them as row vectors. Consequently, after getting data out of a
file, it must be transposed. Data that leaves the chemo functions
(either to db or to gui) is to be transposed again for cleanness.